Housing Values in Suburbs of Boston

Housing Values in Suburbs of Boston

Boston

Format

A data frame with 506 rows and 14 variables:

crim

per capita crime rate by town.

zn

proportion of residential land zoned for lots over 25,000 sq.ft.

indus

proportion of non-retail business acres per town.

chas

Charles River dummy variable (= 1 if tract bounds river; 0 otherwise).

nox

nitrogen oxides concentration (parts per 10 million).

rm

average number of rooms per dwelling.

age

proportion of owner-occupied units built prior to 1940.

dis

weighted mean of distances to five Boston employment centres.

rad

index of accessibility to radial highways.

tax

full-value property-tax rate per $10,000.

ptratio

pupil-teacher ratio by town.

black

proportion of blacks by town.

lstat

lower status of the population (percent).

medv

median value of owner-occupied homes in $1000s.

Source

Harrison, D. and Rubinfeld, D.L. (1978) Hedonic prices and the demand for clean air. J. Environ. Economics and Management 5, 81–102.

Belsley D.A., Kuh, E. and Welsch, R.E. (1980) Regression Diagnostics. Identifying Influential Data and Sources of Collinearity. New York: Wiley.

Details

The Boston data frame was obtained from Venables and Ripley's MASS package. The data utilize census tracts in the Boston Standard Metropolitatn Statistical Area in 1970. Two changes have been made from this original dataset. The dollar values for tax and medv have been converted to 2020 US dollars (assuming a 299.4% cumulative inflation rate). Additionally, the black variable has been transformed from its original metric (1000*(proportion of blacks by town - 0.63)^2) to a simple the proportion of blacks by town.

Examples

summary(Boston)
#>       crim                zn             indus            chas        
#>  Min.   : 0.00632   Min.   :  0.00   Min.   : 0.46   Min.   :0.00000  
#>  1st Qu.: 0.08205   1st Qu.:  0.00   1st Qu.: 5.19   1st Qu.:0.00000  
#>  Median : 0.25651   Median :  0.00   Median : 9.69   Median :0.00000  
#>  Mean   : 3.61352   Mean   : 11.36   Mean   :11.14   Mean   :0.06917  
#>  3rd Qu.: 3.67708   3rd Qu.: 12.50   3rd Qu.:18.10   3rd Qu.:0.00000  
#>  Max.   :88.97620   Max.   :100.00   Max.   :27.74   Max.   :1.00000  
#>       nox               rm             age              dis        
#>  Min.   :0.3850   Min.   :3.561   Min.   :  2.90   Min.   : 1.130  
#>  1st Qu.:0.4490   1st Qu.:5.886   1st Qu.: 45.02   1st Qu.: 2.100  
#>  Median :0.5380   Median :6.208   Median : 77.50   Median : 3.207  
#>  Mean   :0.5547   Mean   :6.285   Mean   : 68.57   Mean   : 3.795  
#>  3rd Qu.:0.6240   3rd Qu.:6.623   3rd Qu.: 94.08   3rd Qu.: 5.188  
#>  Max.   :0.8710   Max.   :8.780   Max.   :100.00   Max.   :12.127  
#>       rad              tax            ptratio          black       
#>  Min.   : 1.000   Min.   : 746.1   Min.   :12.60   Min.   :0.6479  
#>  1st Qu.: 4.000   1st Qu.:1113.2   1st Qu.:17.40   1st Qu.:1.2427  
#>  Median : 5.000   Median :1316.7   Median :19.05   Median :1.2557  
#>  Mean   : 9.549   Mean   :1628.9   Mean   :18.46   Mean   :1.2163  
#>  3rd Qu.:24.000   3rd Qu.:2657.3   3rd Qu.:20.20   3rd Qu.:1.2595  
#>  Max.   :24.000   Max.   :2836.9   Max.   :22.00   Max.   :1.2600  
#>      lstat            medv       
#>  Min.   : 1.73   Min.   : 19.95  
#>  1st Qu.: 6.95   1st Qu.: 67.93  
#>  Median :11.36   Median : 84.59  
#>  Mean   :12.65   Mean   : 89.91  
#>  3rd Qu.:16.95   3rd Qu.: 99.75  
#>  Max.   :37.97   Max.   :199.50