Forest fires in Brazil

This dataset reports the number of forest fires in Brazil per state. The series has data from a 10 year period from 1998 to 2017.

Format

A data frame with 6454 rows and 5 variables:

id

ID number

year

Year when forest fire(s) occured

state

Name of the state of Brazil where the forest fire(s) occured

month

Month when the forest fire(s)

number

Number of forest fires reported

date

Date when forest fire(s) records were made

Source

Sistema Nacional de Informacoes Florestais

Examples

summary(amazon)
#>        id            year         state              month          
#>  Min.   :   1   Min.   :1998   Length:6454        Length:6454       
#>  1st Qu.:1614   1st Qu.:2002   Class :character   Class :character  
#>  Median :3228   Median :2007   Mode  :character   Mode  :character  
#>  Mean   :3228   Mean   :2007                                        
#>  3rd Qu.:4841   3rd Qu.:2012                                        
#>  Max.   :6454   Max.   :2017                                        
#>      number           date           
#>  Min.   :  0.0   Min.   :1998-01-01  
#>  1st Qu.:  3.0   1st Qu.:2002-01-01  
#>  Median : 24.0   Median :2007-01-01  
#>  Mean   :108.3   Mean   :2007-06-18  
#>  3rd Qu.:113.0   3rd Qu.:2012-01-01  
#>  Max.   :998.0   Max.   :2017-01-01