Seoul Bike Sharing Demand

Seoul Bike Sharing Demand

bikes

Format

A data frame with 8760 rows and 14 variables:

Date

mm/dd/yyyy.

Rented

count of bikes rented at each hour.

Hour

hour of he day.

Temp

temperature in Celsius.

Humidity

humidity in percent.

Wind

wind speed in meters/sec.

Visibility

visibility at 10 meters.

Dew

dew point temperature in celcius.

Solar

solar radiation in Mj/m2.

Rainfall

rainfall in millimeters.

Snowfall

snowfall in centimeters.

Seasons

"Winter", "Spring", "Summer", "Autumn".

Holiday

"Holiday", "No Holiday".

Functioning_Day

Does the rental bike system operate on that day?"Yes", "No" .

Source

Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science. Seoul Bike Sharing Demand Data Set

Details

The dataset contains the number of public bikes rented at each hour from the Seoul Bike Sharing System with the corresponding Weather and holiday information. Date is recorded as a character variable. It is a good dataset for feature engineering with dates.

Examples

summary(bikes)

#>      Date               Rented            Hour            Temp       
#>  Length:8760        Min.   :   0.0   Min.   : 0.00   Min.   :-17.80  
#>  Class :character   1st Qu.: 191.0   1st Qu.: 5.75   1st Qu.:  3.50  
#>  Mode  :character   Median : 504.5   Median :11.50   Median : 13.70  
#>                     Mean   : 704.6   Mean   :11.50   Mean   : 12.88  
#>                     3rd Qu.:1065.2   3rd Qu.:17.25   3rd Qu.: 22.50  
#>                     Max.   :3556.0   Max.   :23.00   Max.   : 39.40  
#>     Humidity          Wind         Visibility        Dew         
#>  Min.   : 0.00   Min.   :0.000   Min.   :  27   Min.   :-30.600  
#>  1st Qu.:42.00   1st Qu.:0.900   1st Qu.: 940   1st Qu.: -4.700  
#>  Median :57.00   Median :1.500   Median :1698   Median :  5.100  
#>  Mean   :58.23   Mean   :1.725   Mean   :1437   Mean   :  4.074  
#>  3rd Qu.:74.00   3rd Qu.:2.300   3rd Qu.:2000   3rd Qu.: 14.800  
#>  Max.   :98.00   Max.   :7.400   Max.   :2000   Max.   : 27.200  
#>      Solar           Rainfall          Snowfall         Seasons         
#>  Min.   :0.0000   Min.   : 0.0000   Min.   :0.00000   Length:8760       
#>  1st Qu.:0.0000   1st Qu.: 0.0000   1st Qu.:0.00000   Class :character  
#>  Median :0.0100   Median : 0.0000   Median :0.00000   Mode  :character  
#>  Mean   :0.5691   Mean   : 0.1487   Mean   :0.07507                     
#>  3rd Qu.:0.9300   3rd Qu.: 0.0000   3rd Qu.:0.00000                     
#>  Max.   :3.5200   Max.   :35.0000   Max.   :8.80000                     
#>    Holiday          Functioning_Day   
#>  Length:8760        Length:8760       
#>  Class :character   Class :character  
#>  Mode  :character   Mode  :character  
#>                                       
#>                                       
#>