Coffee Quality Institute Database

Data scraped from the Coffee Quality Institute Database of 1312 arabica coffee beans. Contains both data about the production and taste of the beans. Scores are aggregated based on multiple cups of coffee with many reviewers.

coffee

Format

A data frame with 1311 rows and 28 variables. The variables are as follows:

country_of_origin

Country of Origin of Beans

company

Company which made the coffee

number_of_bags

Number of bags harvested

in_country_partner

Partner of the Company in the Country of Origin

grading_date

Date the Coffee was reviewed

owner_1

Owner of the Company or the Company Name

variety

Variety of Coffee

processing_method

Method by Which the Beans were Processed

aroma

Reviewer's Score of Aroma on a 1-10 scale

flavor

Reviewer's Score of Flavor on a 1-10 scale

aftertaste

Reviewer's Score of Aftertaste on a 1-10 scale

acidity

Reviewer's Score of Acidity on a 1-10 scale

body

Reviewer's Score of Body on a 1-10 scale

balance

Reviewer's Score of Balance on a 1-10 scale

uniformity

Reviewer's Score of Uniformity on a 1-10 scale

clean_cup

Reviewer's Score of "Transparency of a cup" on a 1-10 scale

sweetness

Reviewer's Score of Sweetness on a 1-10 scale

cupper_points

Reviewer's Holistic Score of the cup on a 1-10 scale

total_cup_points

Sum of the Reviewer's Scores

moisture

Moisture percentage from Green Analysis

category_one_defects

Major defects in the beans

color

Color of the Greens

category_two_defects

Minor defects in the beans

expiration

Expiration of the certification of the bean

unit_of_measurement

Unit of Measurement for the Altitude of the farm

altitude_mean_meters

Altitude of the farm

Source

The data is a subset scraped from the Coffee Quality Institute website using the scraper developed by *jldbc*. Full data can be found at https://github.com/jldbc/coffee-quality-database.

Examples

summary(coffee)
#>               country_of_origin   company          number_of_bags  
#>  Mexico                :236     Length:1311        Min.   :   0.0  
#>  Colombia              :183     Class :character   1st Qu.:  14.5  
#>  Guatemala             :181     Mode  :character   Median : 175.0  
#>  Brazil                :132                        Mean   : 153.9  
#>  Taiwan                : 75                        3rd Qu.: 275.0  
#>  United States (Hawaii): 73                        Max.   :1062.0  
#>  (Other)               :431                                        
#>  in_country_partner  grading_date          owner_1             variety   
#>  Length:1311        Min.   :2010-04-09   Length:1311        Caturra:256  
#>  Class :character   1st Qu.:2012-08-01   Class :character   Bourbon:226  
#>  Mode  :character   Median :2014-03-20   Mode  :character   Typica :211  
#>                     Mean   :2014-03-18                      Other  :108  
#>                     3rd Qu.:2015-07-17                      Catuai : 74  
#>                     Max.   :2018-01-19                      (Other):235  
#>                                                             NA's   :201  
#>                  processing_method     aroma           flavor     
#>  Natural / Dry            :251     Min.   :0.000   Min.   :0.000  
#>  Other                    : 26     1st Qu.:7.420   1st Qu.:7.330  
#>  Pulped natural / honey   : 14     Median :7.580   Median :7.580  
#>  Semi-washed / Semi-pulped: 56     Mean   :7.564   Mean   :7.518  
#>  Washed / Wet             :812     3rd Qu.:7.750   3rd Qu.:7.750  
#>  NA's                     :152     Max.   :8.750   Max.   :8.830  
#>                                                                   
#>    aftertaste       acidity           body          balance     
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:7.250   1st Qu.:7.330   1st Qu.:7.330   1st Qu.:7.330  
#>  Median :7.420   Median :7.500   Median :7.500   Median :7.500  
#>  Mean   :7.398   Mean   :7.533   Mean   :7.518   Mean   :7.518  
#>  3rd Qu.:7.580   3rd Qu.:7.750   3rd Qu.:7.670   3rd Qu.:7.750  
#>  Max.   :8.670   Max.   :8.750   Max.   :8.580   Max.   :8.750  
#>                                                                 
#>    uniformity       clean_cup        sweetness      cupper_points   
#>  Min.   : 0.000   Min.   : 0.000   Min.   : 0.000   Min.   : 0.000  
#>  1st Qu.:10.000   1st Qu.:10.000   1st Qu.:10.000   1st Qu.: 7.250  
#>  Median :10.000   Median :10.000   Median :10.000   Median : 7.500  
#>  Mean   : 9.833   Mean   : 9.833   Mean   : 9.903   Mean   : 7.498  
#>  3rd Qu.:10.000   3rd Qu.:10.000   3rd Qu.:10.000   3rd Qu.: 7.750  
#>  Max.   :10.000   Max.   :10.000   Max.   :10.000   Max.   :10.000  
#>                                                                     
#>  total_cup_points    moisture       category_one_defects          color    
#>  Min.   : 0.00    Min.   :0.00000   Min.   : 0.0000      Blue-Green  : 82  
#>  1st Qu.:81.17    1st Qu.:0.09000   1st Qu.: 0.0000      Bluish-Green:112  
#>  Median :82.50    Median :0.11000   Median : 0.0000      Green       :850  
#>  Mean   :82.12    Mean   :0.08886   Mean   : 0.4264      None        : 51  
#>  3rd Qu.:83.67    3rd Qu.:0.12000   3rd Qu.: 0.0000      NA's        :216  
#>  Max.   :90.58    Max.   :0.28000   Max.   :31.0000                        
#>                                                                            
#>  category_two_defects   expiration         unit_of_measurement
#>  Min.   : 0.000       Min.   :2011-04-09   ft: 182            
#>  1st Qu.: 0.000       1st Qu.:2013-08-01   m :1129            
#>  Median : 2.000       Median :2015-03-20                      
#>  Mean   : 3.592       Mean   :2015-03-18                      
#>  3rd Qu.: 4.000       3rd Qu.:2016-07-16                      
#>  Max.   :55.000       Max.   :2019-01-19                      
#>                                                               
#>  altitude_mean_meters
#>  Min.   :     1      
#>  1st Qu.:  1100      
#>  Median :  1311      
#>  Mean   :  1784      
#>  3rd Qu.:  1600      
#>  Max.   :190164      
#>  NA's   :227