Montgomery County, Maryland car crash data

This dataset provides general information about each collision and details of all traffic collisions occurring on county and local roadways within Montgomery County between January 2014 and October 2018, as collected via the Automated Crash Reporting System (ACRS) of the Maryland State Police, and reported by the Montgomery County Police, Gaithersburg Police, Rockville Police, or the Maryland-National Capital Park Police.

Format

A data frame with 44379 rows and 23 variables:

Crash_Type

Identifies crash as property, injury, or fatal.

Route_Type

Type of roadway at crash location.

Cross_Street_Type

Roadway type for nearest cross-street.

Municipality

Jurisdiction for crash location.

Related_NonMotorist

Type(s) of non-motorist involved.

At_Fault

Fault by a driver or non-motorist.

Collision_Type

Type of collision.

Weather

Weather at collision location.

Surface_Condition

Condition of roadway surface.

Brightness

Lighting conditions.

Traffic_Control

Signage or traffic control devices.

Driver_Substance_Abuse

Substance abuse detected for all drivers involved.

NonMotorist_Substance_Abuse

Substance abuse detected for all non-motorists involved.

First_Collision_Event

The first event of the collision.

Second_Collision_Event

Notes is a second collision event took place.

Fixed_Object_Struck

The fixed object struck by vehicle (if applicable).

Junction

The type of junction where the collision occurred.

Intersection_Type

If the collision was intersection related, this field describes the intersection characteristics.

Road_Alignment

The road alignment where the collision occurred.

Road_Condition

The condition of the road when the collision occurred.

Road_Division

The road division where the collision occurred.

Latitude

Y coordinate of crash location.

Longitude

X coordinate of crash location.

Source

Montgomery Car Crash Data

Examples

summary(crashes)
#>     Crash_Type                   Route_Type               Cross_Street_Type
#>  Nonfatal:44270   County              :19022   County              :26556  
#>  Fatal   :  109   Maryland (State)    :18456   Maryland (State)    : 6627  
#>                   Municipality        : 2607   Municipality        : 4878  
#>                   US (State)          : 1954   Ramp                : 2313  
#>                   Interstate (State)  : 1189   Other Public Roadway: 1661  
#>                   Other Public Roadway:  606   Government          :  884  
#>                   (Other)             :  545   (Other)             : 1460  
#>            Municipality               Related_NonMotorist        At_Fault    
#>  GAITHERSBURG    :15889                         :    0    BOTH       :  126  
#>  ROCKVILLE       :14004   BICYCLIST/PEDALCYCLIST:17190    DRIVER     :43544  
#>  WASHINGTON GROVE: 2516   OTHER                 :10443    NONMOTORIST:  709  
#>  GARRETT PARK    : 1508   PEDESTRIAN            :16746    UNKNOWN    :    0  
#>  KENSINGTON      : 1217                                                      
#>  CHEVY CHASE #5  : 1189                                                      
#>  (Other)         : 8056                                                      
#>             Collision_Type                   Weather     
#>  Angle             : 7358   CLEAR                :31866  
#>  HEAD ON           : 3935   CLOUDY               : 5218  
#>  Opposite Direction:  854   FOGGY                :  298  
#>  OTHER             : 6087   OTHER                :  184  
#>  Same Direction    :18818   RAINING              : 5788  
#>  SINGLE VEHICLE    : 7327   Snow/Sleet/Wintry Mix:  863  
#>                             Windy                :  162  
#>       Surface_Condition          Brightness                   Traffic_Control 
#>  DRY           :34288   DARK          : 2195   NO CONTROLS            :23765  
#>  Ice/Snow/Slush: 1452   DARK LIGHTS ON:10108   TRAFFIC SIGNAL         :14217  
#>  Other         :  186   Dawn/Dusk     : 2036   STOP SIGN              : 3493  
#>  WET           : 8453   DAYLIGHT      :29778   OTHER                  :  935  
#>                         OTHER         :  262   FLASHING TRAFFIC SIGNAL:  641  
#>                                                YIELD SIGN             :  596  
#>                                                (Other)                :  732  
#>    Driver_Substance_Abuse NonMotorist_Substance_Abuse
#>  Alcohol      : 4112      Alcohol      :  575        
#>  Combination  :  434      Illegal Drug :    7        
#>  Illegal Drug :  650      Medication   :    2        
#>  Medication   :  974      NONE DETECTED:43795        
#>  NONE DETECTED:38110                                 
#>  OTHER        :   99                                 
#>  UNKNOWN      :    0                                 
#>         First_Collision_Event Second_Collision_Event        Fixed_Object_Struck
#>  OTHER VEHICLE     :30630     No :33426              Bridge           :  253   
#>  FIXED OBJECT      : 5464     Yes:10953              BUILDING         : 1248   
#>  PARKED VEHICLE    : 4336                            Curb/Ditch       :24487   
#>  PEDESTRIAN        : 1684                            FENCE            : 1945   
#>  OFF ROAD          :  505                            Guardrail/Barrier: 3661   
#>  Bicycle/Pedalcycle:  454                            Other            :10980   
#>  (Other)           : 1306                            Tree             : 1805   
#>              Junction                 Intersection_Type  Road_Alignment 
#>  DRIVEWAY        : 1543   FIVE-POINT OR MORE   :  304   Curve   : 6468  
#>  Intersection    :22590   FOUR-WAY INTERSECTION:19259   OTHER   : 1262  
#>  NON INTERSECTION:18595   OTHER                : 8327   STRAIGHT:36649  
#>  Other           : 1651   ROUNDABOUT           :  311                   
#>                           T-INTERSECTION       :15203                   
#>                           TRAFFIC CIRCLE       :   79                   
#>                           Y-INTERSECTION       :  896                   
#>  Road_Condition Road_Division      Latitude       Longitude     
#>  No :43398      One-way: 1419   Min.   :38.01   Min.   :-79.43  
#>  Yes:  981      OTHER  : 1748   1st Qu.:39.03   1st Qu.:-77.19  
#>                 Two-way:41212   Median :39.08   Median :-77.11  
#>                                 Mean   :39.08   Mean   :-77.12  
#>                                 3rd Qu.:39.14   3rd Qu.:-77.04  
#>                                 Max.   :39.99   Max.   :-75.53  
#>