FiveThirtyEight hate crimes dataset

This dataset contains the data that supports the article, "Higher Rates of Hate Crimes Are Tied to Income Inequality", published by FiveThirtyEight on 23 January, 2017, authored by Maimuna Majumder.

hate_crimes

Format

A data frame with 51 rows and 12 variables:

state

character State name

median_household_income

integer Median household income, 2016

share_unemployed_seasonal

double Share of the population that is unemployed

share_population_in_metro_areas

double Share of the population that lives in metropolitan areas, 2015

share_population_with_high_school_degree

double Share of adults 25 and older with a high-school degree, 2009

share_non_citizen

double Share of the population that are not U.S. citizens, 2015

share_white_poverty

double Share of white residents who are living in poverty, 2015

gini_index

double Gini Index, 2015

share_non_white

double Share of the population that is not white, 2015

share_voters_voted_trump

double Share of 2016 U.S. presidential voters who voted for Donald Trump

hate_crimes_per_100k_splc

double Hate crimes per 100,000 population, Southern Poverty Law Center, Nov. 9-18, 2016

avg_hatecrimes_per_100k_fbi

double Average annual hate crimes per 100,000 population, FBI, 2010-2015

Source

Fivethirtyeight Kaggle repository

Examples

summary(hate_crimes)
#>     state           median_household_income share_unemployed_seasonal
#>  Length:51          Min.   :35521           Min.   :0.02800          
#>  Class :character   1st Qu.:48657           1st Qu.:0.04200          
#>  Mode  :character   Median :54916           Median :0.05100          
#>                     Mean   :55224           Mean   :0.04957          
#>                     3rd Qu.:60719           3rd Qu.:0.05750          
#>                     Max.   :76165           Max.   :0.07300          
#>                                                                      
#>  share_population_in_metro_areas share_population_with_high_school_degree
#>  Min.   :0.3100                  Min.   :0.7990                          
#>  1st Qu.:0.6300                  1st Qu.:0.8405                          
#>  Median :0.7900                  Median :0.8740                          
#>  Mean   :0.7502                  Mean   :0.8691                          
#>  3rd Qu.:0.8950                  3rd Qu.:0.8980                          
#>  Max.   :1.0000                  Max.   :0.9180                          
#>                                                                          
#>  share_non_citizen share_white_poverty   gini_index     share_non_white 
#>  Min.   :0.01000   Min.   :0.04000     Min.   :0.4190   Min.   :0.0600  
#>  1st Qu.:0.03000   1st Qu.:0.07500     1st Qu.:0.4400   1st Qu.:0.1950  
#>  Median :0.04500   Median :0.09000     Median :0.4540   Median :0.2800  
#>  Mean   :0.05458   Mean   :0.09176     Mean   :0.4538   Mean   :0.3157  
#>  3rd Qu.:0.08000   3rd Qu.:0.10000     3rd Qu.:0.4665   3rd Qu.:0.4200  
#>  Max.   :0.13000   Max.   :0.17000     Max.   :0.5320   Max.   :0.8100  
#>  NA's   :3                                                              
#>  share_voters_voted_trump hate_crimes_per_100k_splc avg_hatecrimes_per_100k_fbi
#>  Min.   :0.040            Min.   :0.06745           Min.   : 0.2669            
#>  1st Qu.:0.415            1st Qu.:0.14271           1st Qu.: 1.2931            
#>  Median :0.490            Median :0.22620           Median : 1.9871            
#>  Mean   :0.490            Mean   :0.30409           Mean   : 2.3676            
#>  3rd Qu.:0.575            3rd Qu.:0.35694           3rd Qu.: 3.1843            
#>  Max.   :0.700            Max.   :1.52230           Max.   :10.9535            
#>                           NA's   :4                 NA's   :1