Time Spent Watching TV on a Day Surveyed, 2017

This is a data set detailing TV usage on days surveyed as determined by the 2017 American Time Use Survey. The data set includes demographic information, as well as details regarding employment and family makeup, where applicable. Information on days surveyed, as well as whether the day is a holiday, is also included.

tv17

Arguments

id
ID of respondent

weight
ATUS final weight

youngest_child
Age of the youngest child in the household that is less than 18 years old (if applicable). Range: 1-17; if no child in household: NA

age
Age of respondent

sex
Sex of respondent

job
Status of employment of the respondent. Direct transcription from original codebook:

"1"Employed, at work

"2"Employed, absent

"3"Unemployed, on layoff

"4"Unemployed, looking

"5"Not in the labor force

m_job
The response to question, “in the last seven days did you have more than one job?” Returns NA if no job.

f_job
Does the respondent have a full time job or a part time job? (NA if no job)

educ
Are you enrolled in high school, college, or university? (NA if not currently enrolled)

educ2
If yes to educ, are you enrolled in high school or upper schooling? (NA if not currently enrolled)

partner
Presence of the respondent’s spouse or unmarried partner in the household

"1"Spouse present

"2"Unmarried partner present

"3"No spouse/unmarried partner present

pr_job
Answer to the question, “does your partner have a job?” (NA if not applicable)

salary
Weekly earnings at the respondent’s main job, two decimals implied

children
Number of children under 18 in the household

pr_job_f
Part time/full time job status of partner, if applicable (NA if partner unemployed or no partner)

job_hours
Total hours usually worked per week (-4: Hours vary)

day
Day of the week about which the respondent was interviewed (Monday thorugh Friday)

holiday
Notes if the respondent was interviewed on a holiday

elder_care
Total time spent providing elder care that day by the respondent, in minutes

child_time
Total time spent during diary day providing secondary childcare for household children younger than 13, in minutes

tv
Minutes spent watching TV

Details

For more information regarding the key visit https://www.bls.gov/tus/atusintcodebk17.pdf. This data is retrieved from the American Time Use Survey, made available through the Bureau of Labor Statistics https://www.bls.gov/tus/datafiles_2017.htm.