Coffee Quality Institute Database

Data scraped from the Coffee Quality Institute Database of 1312 arabica coffee beans. Contains both data about the production and taste of the beans. Scores are aggregated based on multiple cups of coffee with many reviewers.

coffee

Format

A data frame with 1311 rows and 28 variables. The variables are as follows:

Country.of.Origin

Country of Origin of Beans

Company

Company which made the coffee

Number.of.Bags

Number of bags harvested

In.Country.Partner

Partner of the Company in the Country of Origin

Grading.Date

Date the Coffee was reviewed

Owner.1

Owner of the Company or the Company Name

Variety

Variety of Coffee

Processing.Method

Method by Which the Beans were Processed

Aroma

Reviewer's Score of Aroma on a 1-10 scale

Flavor

Reviewer's Score of Flavor on a 1-10 scale

Aftertaste

Reviewer's Score of Aftertaste on a 1-10 scale

Acidity

Reviewer's Score of Acidity on a 1-10 scale

Body

Reviewer's Score of Body on a 1-10 scale

Balance

Reviewer's Score of Balance on a 1-10 scale

Uniformity

Reviewer's Score of Uniformity on a 1-10 scale

Clean.Cup

Reviewer's Score of "Transparency of a cup" on a 1-10 scale

Sweetness

Reviewer's Score of Sweetness on a 1-10 scale

Cupper.Points

Reviewer's Holistic Score of the cup on a 1-10 scale

Total.Cup.Points

Sum of the Reviewer's Scores

Moisture

Moisture percentage from Green Analysis

Category.One.Defects

Major defects in the beans

Color

Color of the Greens

Category.Two.Defects

Minor defects in the beans

Expiration

Expiration of the certification of the bean

unit_of_measurement

Unit of Measurement for the Altitude of the farm

altitude_mean_meters

Altitude of the farm

Source

The data is a subset scraped from the Coffee Quality Institute website using the scraper developed by *jldbc*. Full data can be found here.

Examples

summary(coffee)

#>  Country.of.Origin    Company          Number.of.Bags   In.Country.Partner
#>  Length:1311        Length:1311        Min.   :   0.0   Length:1311       
#>  Class :character   Class :character   1st Qu.:  14.5   Class :character  
#>  Mode  :character   Mode  :character   Median : 175.0   Mode  :character  
#>                                        Mean   : 153.9                     
#>                                        3rd Qu.: 275.0                     
#>                                        Max.   :1062.0                     
#>                                                                           
#>   Grading.Date          Owner.1            Variety          Processing.Method 
#>  Min.   :2010-04-09   Length:1311        Length:1311        Length:1311       
#>  1st Qu.:2012-08-01   Class :character   Class :character   Class :character  
#>  Median :2014-03-20   Mode  :character   Mode  :character   Mode  :character  
#>  Mean   :2014-03-18                                                           
#>  3rd Qu.:2015-07-17                                                           
#>  Max.   :2018-01-19                                                           
#>                                                                               
#>      Aroma           Flavor        Aftertaste       Acidity     
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:7.420   1st Qu.:7.330   1st Qu.:7.250   1st Qu.:7.330  
#>  Median :7.580   Median :7.580   Median :7.420   Median :7.500  
#>  Mean   :7.564   Mean   :7.518   Mean   :7.398   Mean   :7.533  
#>  3rd Qu.:7.750   3rd Qu.:7.750   3rd Qu.:7.580   3rd Qu.:7.750  
#>  Max.   :8.750   Max.   :8.830   Max.   :8.670   Max.   :8.750  
#>                                                                 
#>       Body          Balance        Uniformity       Clean.Cup     
#>  Min.   :0.000   Min.   :0.000   Min.   : 0.000   Min.   : 0.000  
#>  1st Qu.:7.330   1st Qu.:7.330   1st Qu.:10.000   1st Qu.:10.000  
#>  Median :7.500   Median :7.500   Median :10.000   Median :10.000  
#>  Mean   :7.518   Mean   :7.518   Mean   : 9.833   Mean   : 9.833  
#>  3rd Qu.:7.670   3rd Qu.:7.750   3rd Qu.:10.000   3rd Qu.:10.000  
#>  Max.   :8.580   Max.   :8.750   Max.   :10.000   Max.   :10.000  
#>                                                                   
#>    Sweetness      Cupper.Points    Total.Cup.Points    Moisture      
#>  Min.   : 0.000   Min.   : 0.000   Min.   : 0.00    Min.   :0.00000  
#>  1st Qu.:10.000   1st Qu.: 7.250   1st Qu.:81.17    1st Qu.:0.09000  
#>  Median :10.000   Median : 7.500   Median :82.50    Median :0.11000  
#>  Mean   : 9.903   Mean   : 7.498   Mean   :82.12    Mean   :0.08886  
#>  3rd Qu.:10.000   3rd Qu.: 7.750   3rd Qu.:83.67    3rd Qu.:0.12000  
#>  Max.   :10.000   Max.   :10.000   Max.   :90.58    Max.   :0.28000  
#>                                                                      
#>  Category.One.Defects    Color           Category.Two.Defects
#>  Min.   : 0.0000      Length:1311        Min.   : 0.000      
#>  1st Qu.: 0.0000      Class :character   1st Qu.: 0.000      
#>  Median : 0.0000      Mode  :character   Median : 2.000      
#>  Mean   : 0.4264                         Mean   : 3.592      
#>  3rd Qu.: 0.0000                         3rd Qu.: 4.000      
#>  Max.   :31.0000                         Max.   :55.000      
#>                                                              
#>    Expiration         unit_of_measurement altitude_mean_meters
#>  Min.   :2011-04-09   Length:1311         Min.   :     1      
#>  1st Qu.:2013-08-01   Class :character    1st Qu.:  1100      
#>  Median :2015-03-20   Mode  :character    Median :  1311      
#>  Mean   :2015-03-18                       Mean   :  1784      
#>  3rd Qu.:2016-07-16                       3rd Qu.:  1600      
#>  Max.   :2019-01-19                       Max.   :190164      
#>                                           NA's   :227