Google Play Store Apps

A data frame containing all the details of the applications on Google Play. There are 13 features that describe a given app.

googleplay

Format

A data frame with 10,842 rows and 13 variables The variables are as follows:

App

application name

Category

category the app belongs to ss

Rating

overall user rating of the app

Reviews

number of user reviews for the app

Size

size of the app

Installs

number of user downloads/installs for the app

Type

paid or free

Price

price of the app

Content Rating

age group the app is targeted at

Genres

an app can belong to multiple genres apart from its main category

Last Updated

date when the app was last updated on play store

Current Ver

current version of the app available on play store

Android Ver

min required android version

Source

The data is from Kaggle. Lavanya Gupta created the Google Play Store Apps on Kaggle. https://www.kaggle.com/lava18/google-play-store-apps

Details

Lavanya Gupta created a public dataset on Kaggle about Google Play Store applications. The information in the dataset was scraped from the Google Play Store. Because there are apps always being added or removed from Google Play Store, Gupta has modified this dataset over time. This dataset is the most recent version, Version 5, and was uploaded by Gupta to Kaggle on September 2018.

Examples

table(googleplay$Category,googleplay$Type)

#>                      
#>                       Free Paid
#>   ART_AND_DESIGN        62    3
#>   AUTO_AND_VEHICLES     82    3
#>   BEAUTY                53    0
#>   BOOKS_AND_REFERENCE  203   28
#>   BUSINESS             446   14
#>   COMICS                60    0
#>   COMMUNICATION        360   27
#>   DATING               227    7
#>   EDUCATION            152    4
#>   ENTERTAINMENT        147    2
#>   EVENTS                63    1
#>   FAMILY              1780  191
#>   FINANCE              349   17
#>   FOOD_AND_DRINK       125    2
#>   GAME                1061   83
#>   HEALTH_AND_FITNESS   325   16
#>   HOUSE_AND_HOME        88    0
#>   LIBRARIES_AND_DEMO    84    1
#>   LIFESTYLE            363   19
#>   MAPS_AND_NAVIGATION  132    5
#>   MEDICAL              354  109
#>   NEWS_AND_MAGAZINES   281    2
#>   PARENTING             58    2
#>   PERSONALIZATION      309   83
#>   PHOTOGRAPHY          313   22
#>   PRODUCTIVITY         396   28
#>   SHOPPING             258    2
#>   SOCIAL               292    3
#>   SPORTS               360   24
#>   TOOLS                765   78
#>   TRAVEL_AND_LOCAL     246   12
#>   VIDEO_PLAYERS        171    4
#>   WEATHER               74    8